requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.9
botocore==1.38.9
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2025.4
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-tables2==2.7.5
django_celery_results==2.6.0
djangorestframework==3.16.0
docopt==0.6.2
dotenv==0.9.9
ffmpeg-python==0.2.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
GPUtil==1.4.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kafka-python==2.2.3
kombu==5.5.3
MarkupSafe==3.0.2
nltk==3.9.1
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.5
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pexels==1.1
pythondialog==3.5.3
redis==6.0.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
